Add files via upload
""And no matter how lazy you think they are, boredom is a powerful motivator, and so is a desire not to let down your team, and so is a desire not to look bad in
front of others: bring 'em to work, leave 'em alone, and nearly all of them actually will work, will actually build things that are built
well, built for the ages, built to last. Paradoxically, the really wasted money is the money that gets spent on government overseers
determined to make sure that none of the workers waste any money: point people at jobs, give 'em simple hand tools, and tell them to
take their time and build something solid and it's almost impossible for us to not get that money back in long-term savings."  https://web.archive.org/web/20111103112736/http://bradhicks.livejournal.com/422902.html 

https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3603114
p.10
"F. Dignity Under the 1987 Constitution The 40-year judicial discourse on the egalitarian conception of dignity may have helped pave the way for the express recognition of the right of every person to human dignity in the present 1987 Constitution. 51 For the framers of the 1987 Constitution, the protection of human dignity meant the promotion and protection of a whole gamut of rights “pertinent to the existence of the human person” and those that they may eventually discover.52 Human rights were not to be considered the end in themselves but, rather, “as means to achieve greater ends.”53 As a result, the country’s current 1987 Constitution ensured the “supremacy of human dignity over things”54 so much so that it dedicated the entire Article XIII, composed of 16 sections, to detail various forms of social justice reforms that specifically tackles labor, agrarian and natural resources, urban land and housing, health, women, and indigenous peoples. Crucially, it established a Commission on Human Rights for the promotion and protection of human dignity. 55 Some of the important cases decided by the Court under the 1987 Constitution will be discussed in Part III of this chapter" https://www.rappler.com/voices/imho/243165-no-human-dignity-philippines/  Labor, Unemployment, another constitution, the 1935 Works Progress Administration, and building buildings that last are something only an apophenic would connect.